+++ packages/test-sync/src/commonTest/kotlin/io/realm/kotlin/test/mongodb/common/AppTests.kt
@@ -523,7 +523,7 @@ class AppTests {
         ).use { testApp ->
             assertEquals(SyncServerConfig.url, testApp.baseUrl)
 
-            RealmLog.level = LogLevel.ALL
+            RealmLog.setLevel(LogLevel.ALL)
             runBlocking {
                 testApp.updateBaseUrl(null)
             }