+++ packages/test-sync/src/commonTest/kotlin/io/realm/kotlin/test/mongodb/common/HttpLogObfuscatorTests.kt
@@ -13,11 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+@file:Suppress("invisible_reference", "invisible_member")
 
 package io.realm.kotlin.test.mongodb.common
 
 import io.realm.kotlin.internal.platform.runBlocking
+import io.realm.kotlin.log.LogCategory
 import io.realm.kotlin.log.LogLevel
+import io.realm.kotlin.log.RealmLog
 import io.realm.kotlin.log.RealmLogger
 import io.realm.kotlin.mongodb.Credentials
 import io.realm.kotlin.mongodb.GoogleAuthType
@@ -52,11 +55,8 @@ class HttpLogObfuscatorTests {
     private class ObfuscatorLoggerInspector(
         private val channel: Channel<Operation>
     ) : RealmLogger {
-
-        override val level: LogLevel = LogLevel.DEBUG
-        override val tag: String = "ObfuscatorLoggerInspector"
-
         override fun log(
+            category: LogCategory,
             level: LogLevel,
             throwable: Throwable?,
             message: String?,
@@ -114,15 +114,15 @@ class HttpLogObfuscatorTests {
 
     @BeforeTest
     fun setUp() {
+        RealmLog.setLevel(LogLevel.DEBUG)
         channel = Channel(1)
+        RealmLog.add(ObfuscatorLoggerInspector(channel))
     }
 
     private fun initApp(): TestApp {
         return TestApp(
             this::class.simpleName,
             appName = syncServerAppName("obfsctr"),
-            logLevel = LogLevel.DEBUG,
-            customLogger = ObfuscatorLoggerInspector(channel),
             initialSetup = { app, service ->
                 initializeDefault(app, service)
                 app.addFunction(TestAppInitializer.FIRST_ARG_FUNCTION)
@@ -135,6 +135,7 @@ class HttpLogObfuscatorTests {
     @AfterTest
     fun tearDown() {
         channel.close()
+        RealmLog.reset()
 
         if (this::app.isInitialized) {
             app.close()
@@ -143,13 +144,12 @@ class HttpLogObfuscatorTests {
 
     @Test
     fun nullObfuscator() = runBlocking {
-        val logger = CustomLogCollector("NULL-OBFUSCATOR", LogLevel.DEBUG)
+        val logger = CustomLogCollector()
+        RealmLog.add(logger)
         app = TestApp(
             "nullObfuscator",
             appName = syncServerAppName("null-obf"),
-            logLevel = LogLevel.DEBUG,
             builder = { it.httpLogObfuscator(null) },
-            customLogger = logger,
             initialSetup = { app, service ->
                 initializeDefault(app, service)
                 app.addFunction(TestAppInitializer.FIRST_ARG_FUNCTION)